refactor(smtp_server): splitter en headers/commands + auth étendu (689→378 LOC)
